% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coef.r
\name{coef.crm}
\alias{coef.crm}
\title{Extract coefficients}
\usage{
\method{coef}{crm}(object,...)
}
\arguments{
\item{object}{crm model result}

\item{...}{generic arguments not used here}
}
\value{
returns a dataframe with estimates and standard
errors and confidence intervals if hessian=TRUE on model run.
}
\description{
Extracts the beta coefficients from the model results.
}
\author{
Jeff Laake
}
\seealso{
\code{\link{crm}}
}
\keyword{utility}

