% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_matrices}
\alias{compute_matrices}
\title{Compute HMM matrices}
\usage{
compute_matrices(object, ddl = NULL)
}
\arguments{
\item{object}{fitted crm model (must be an HMM model)}

\item{ddl}{design data list}
}
\value{
list with gamma, dmat and delta arrays
}
\description{
Computes gamma, dmat and delta (initial) matrices(arrays) and returns them in a list.
}
\author{
Jeff Laake
}
\keyword{utility}

