% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.fixed.matrix}
\alias{create.fixed.matrix}
\title{Create parameters with fixed matrix}
\usage{
create.fixed.matrix(ddl, parameters)
}
\arguments{
\item{ddl}{design data list}

\item{parameters}{parameter specififcation list}
}
\value{
parameters with fixed matrix set
}
\description{
Creates fixed matrix in parameters from ddl$fix values
}
\author{
Jeff Laake
}
\keyword{utility}

