% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm2hmm.r
\name{dmat_hsmm2hmm}
\alias{dmat_hsmm2hmm}
\title{Create expanded state-dependent observation matrix for HMM from HSMM}
\usage{
dmat_hsmm2hmm(dmat, mv)
}
\arguments{
\item{dmat}{state-dependent observation matrix for aggregated states}

\item{mv}{vector of dwell time distribution lengths}
}
\value{
expanded state-dependent observation matrix for hmm to approximate hsmm
}
\description{
Creates expanded state-dependent matrix for HMM from aggregated state-dependent observation
matrix.
}
\author{
Jeff Laake
}
\keyword{utility}

