% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm2hmm.r
\name{hsmm2hmm}
\alias{hsmm2hmm}
\title{Compute transition matrix for HMM from HSMM}
\usage{
hsmm2hmm(omega, dm, eps = 1e-10)
}
\arguments{
\item{omega}{transition matrix for aggregated states}

\item{dm}{list of dwell time distribution vectors; each list element is vector of dwell time probabilities in order of aggregated states}

\item{eps}{epsilon value for setting small probabilities to 0; anything less than eps is set to exactly 0.}
}
\value{
transition matrix for hmm to approximate hsmm
}
\description{
Computes expanded transition matrix for HMM from aggregated state transition
matrix and dwell time distributions.
}
\author{
Walter Zucchini
}
\keyword{utility}

