% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mvms.r
\name{mvms_design_data}
\alias{mvms_design_data}
\title{Multivariate Multistate (mvms) Design Data}
\usage{
mvms_design_data(df.states, df = NULL, transition = TRUE)
}
\arguments{
\item{df.states}{dataframe of states created from set_mvms}

\item{df}{is dataframe of observations; provided for parameter delta in which both state and observation are needed in design data}

\item{transition}{if TRUE, creates design data for a state transition (from to); otherwise just state variables}
}
\value{
a dataframe to be used with design data for mvms model
}
\description{
Creates a dataframe with design data for MvMS model for a single occasion
}
\examples{
x=set_mvms(list(location=c("A","B","C"),repro_status=c("N","P","u")))
mvms_design_data(x$df.states)
mvms_design_data(x$df.states,transition=FALSE)

}
\author{
Jeff Laake
}
