% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.crm.r
\name{print.crm}
\alias{print.crm}
\title{Print model results}
\usage{
\method{print}{crm}(x,...)
}
\arguments{
\item{x}{crm model result or list of model results}

\item{...}{generic arguments not used here}
}
\value{
\code{print} prints a simple summary of the model to the screen and
returns NULL.
}
\description{
Provides a printed simple summary of the model results.
}
\author{
Jeff Laake
}
\seealso{
\code{\link{crm}}
}
\keyword{utility}

