% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.crm.r
\name{print.crmlist}
\alias{print.crmlist}
\title{Print model table from model list}
\usage{
\method{print}{crmlist}(x,...)
}
\arguments{
\item{x}{list of model results}

\item{...}{generic arguments not used here}
}
\value{
None
}
\description{
Print model table from model list
}
\author{
Jeff Laake
}
\seealso{
\code{\link{crm}}
}
\keyword{utility}

