% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set.fixed}
\alias{set.fixed}
\title{Set fixed real parameter values in ddl}
\usage{
set.fixed(ddl, parameters)
}
\arguments{
\item{ddl}{design data list}

\item{parameters}{parameter specififcation list}
}
\value{
design data list with fixed values
}
\description{
Merges fixed real parameter values in parameter specification with those specified with fix in the
design data list and returns the fixed values in the design data list.
}
\author{
Jeff Laake
}
\keyword{utility}

