% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admbutils.R
\name{setup_admb}
\alias{setup_admb}
\title{ADMB setup}
\usage{
setup_admb(tpl, compile = FALSE, clean = TRUE, safe = TRUE, re = FALSE)
}
\arguments{
\item{tpl}{character string for admb template file}

\item{compile}{if TRUE forces re-compilation of tpl file}

\item{clean}{if TRUE, deletes the tpl and executable files for amdb in local directory}

\item{safe}{can be used to set safe mode for admb}

\item{re}{uses admb-re if TRUE for random effects}
}
\description{
Sets up executable for the tpl file by looking for exe in package directory or
compiles tpl file in local directory (clean=F) of from package directory.
If admb cannot be found will attempt to run prepare_admb function (if exists) to
establish connections for compilation.
}
