% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admbutils.R
\name{setup_tmb}
\alias{setup_tmb}
\title{TMB setup}
\usage{
setup_tmb(tpl, clean = FALSE, debug = FALSE)
}
\arguments{
\item{tpl}{character string for admb template file}

\item{clean}{if TRUE, deletes the tpl (.cpp) and executable files in local directory and copies from package directory}
}
\description{
Sets up executable for the .cpp file (tpl) by looking for exe in package directory or
compiles cpp file in local directory (clean=FALSE) of from package directory.
}
