% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/js.hessian.R
\name{js.hessian}
\alias{js.hessian}
\title{Compute variance-covariance matrix for fitted JS model}
\usage{
js.hessian(model)
}
\arguments{
\item{model}{fitted JS model from function crm}
}
\value{
variance-covariance matrix for specified model or the model
object with the stored vcv depending on whether the model has already been run
}
\description{
A wrapper function that sets up call to hessian function to compute and then
invert the hessian.
}
\author{
Jeff Laake <jeff.laake@noaa.gov>
}

